rename more-information form to copyright form
