+++ templates/vue/src/components/modals/NodeModal/forms/CopyrightForm.vue
